Added players.py
The players.py file is a structure for the object player to be used in cricket. This file contains different categories of player such as batsman, bowler, wicket keeper and all-rounder.
Various attributes for these player categories, such as runs, overs, wickets are present in this file. Also methods are created for calling and setting this attributes.